+++ src/types/assetType.ts
@@ -13,6 +13,7 @@ export interface CreateAssetType {
   name: string;
   targetPercentage: number;
   assetClassId: number;
+  userId: number;
 }
 
 export interface UpdateAssetType {